status

+++ Src/database/migrations/2020_10_28_041610_create_steps_table.php
@@ -19,8 +19,8 @@ class CreateStepsTable extends Migration
             $table->unsignedBigInteger('approver_id')->comment('User ID');
             $table->tinyInteger('approver_type')->default(0)->comment('0 : TO , 1 : CC');
             $table->tinyInteger('step_type')->default(1)->comment('1 : Application(Step 1) , 2 : Settlement(Step 2)');
-            $table->integer('order')->default(1)->comment('using to update status on applications table when user approved');
-            $table->integer('select_order')->default(1)->comment('using get list approved of user by status');
+            $table->tinyInteger('order')->default(1)->comment('using to update status on applications table when user approved');
+            $table->tinyInteger('select_order')->default(1)->comment('using get list approved of user by status');
             $table->unsignedBigInteger('created_by')->nullable();
             $table->unsignedBigInteger('updated_by')->nullable();
             $table->timestamps();
